2025.05.30
댓글작성, 대댓글작성 테스트코드 구현
redis, mail gradle 추가

+++ src/main/java/com/weatherfit/comment_service/comment/dto/CommentResponseDTO.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Builder
 public class CommentResponseDTO {
     private Long id;
-    private Integer boardId;
+    private Long boardId;
     private String nickname;
     private String content;
